api/requirements.txt
flask
numpy
pandas
python-dotenv
scikit-learn
scipy
supabase
google-generativeai
flask_cors
tensorflow[and-cuda]
ultralytics 
"numpy<2"